Update for 2022

+++ src/main/java/frc/robot/components/Elevator.java
@@ -206,11 +206,11 @@ public class Elevator extends Titan.Component<Robot>{
     }
 
     public int getEncoderPosition(){
-        return bottom.getSelectedSensorPosition();
+        return (int) bottom.getSelectedSensorPosition();
     }
 
     public int getEncoderVelocity(){
-        return bottom.getSelectedSensorVelocity();
+        return (int) bottom.getSelectedSensorVelocity();
     }
 
     public boolean isElevatorDown(){
